backend/backend/requirements.txt
django-filter==22.1
Django==4.1.5
djangorestframework==3.14.0
django-environ==0.9.0
psycopg2==2.9.5
psycopg2-binary==2.9.5
Pillow==9.4.0
fpdf==1.7.2
django-cors-headers==3.13.0
flake8==6.0.0
isort==5.11.4